Split bool match validation
