+++ src/typechecker/patterns/match_validation.rs
@@ -3,10 +3,7 @@ use std::collections::{HashMap, HashSet};
 use crate::ast::expressions::MatchArm;
 use crate::ast::typed::*;
 use crate::ast::Pattern;
-use crate::error::{
-    Diagnostic, Span, SuggestedFix, TextEdit, MISSING_BOOL_MATCH_ARM_FIX_KIND,
-    MISSING_BOOL_MATCH_ARM_FIX_TITLE,
-};
+use crate::error::{Diagnostic, Span};
 
 use super::TypeChecker;
 
@@ -149,84 +146,6 @@ impl TypeChecker {
                 )),
                 _ => {}
             }
-        }
-    }
-
-    pub(crate) fn check_bool_match_patterns(
-        &mut self,
-        arms: &[MatchArm],
-        require_exhaustive: bool,
-        span: Span,
-    ) {
-        let mut true_seen = false;
-        let mut false_seen = false;
-        let mut wildcard_seen = false;
-
-        for arm in arms {
-            match &arm.pattern {
-                Pattern::BoolTrue { span } => {
-                    if true_seen || wildcard_seen {
-                        self.diagnostics.push(Diagnostic::error(
-                            "E4005",
-                            "duplicate match arm for `true`",
-                            *span,
-                        ));
-                    }
-                    true_seen = true;
-                }
-                Pattern::BoolFalse { span } => {
-                    if false_seen || wildcard_seen {
-                        self.diagnostics.push(Diagnostic::error(
-                            "E4005",
-                            "duplicate match arm for `false`",
-                            *span,
-                        ));
-                    }
-                    false_seen = true;
-                }
-                Pattern::Wildcard { span } => {
-                    if wildcard_seen || (true_seen && false_seen) {
-                        self.diagnostics.push(Diagnostic::error(
-                            "E4005",
-                            "redundant wildcard match arm",
-                            *span,
-                        ));
-                    }
-                    wildcard_seen = true;
-                }
-                _ => {}
-            }
-        }
-
-        if require_exhaustive && !wildcard_seen && !(true_seen && false_seen) {
-            let missing_values: Vec<&str> = match (true_seen, false_seen) {
-                (true, false) => vec!["false"],
-                (false, true) => vec!["true"],
-                _ => vec!["true", "false"],
-            };
-            let missing = missing_values
-                .iter()
-                .map(|value| format!("`{value}`"))
-                .collect::<Vec<_>>()
-                .join(", ");
-            let replacement = missing_values
-                .iter()
-                .map(|value| format!("        | {value} {{ <expression> }}"))
-                .collect::<Vec<_>>()
-                .join("\n");
-            let insertion = Span::new(span.file_id, span.end, span.end);
-            self.diagnostics.push(
-                Diagnostic::error(
-                    "E4006",
-                    format!("non-exhaustive bool match: missing {missing}"),
-                    span,
-                )
-                .with_suggested_fix(SuggestedFix::new(
-                    MISSING_BOOL_MATCH_ARM_FIX_KIND,
-                    MISSING_BOOL_MATCH_ARM_FIX_TITLE,
-                    vec![TextEdit::new(insertion, format!("\n{replacement}"))],
-                )),
-            );
         }
     }
